pyproject.toml
[tool.poetry]
name = "easternmind"
version = "0.1.0"
description = ""
authors = ["source64 <65923236+source64@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
black = "^24.8.0"
django-compressor = "^4.5.1"
django-livereload-server = "^0.5.1"
pillow = "^10.4.0"
django-pictures = "^1.3.3"
django-cleanup = "^9.0.0"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
